+++ bindings/react/demo/index.js
@@ -20,6 +20,8 @@ import CardExample from './examples/Card';
 import LazyListExample from './examples/LazyList';
 import TabbarExample from './examples/Tabbar';
 import AlertDialogExample from './examples/AlertDialog';
+import ActionSheetExample from './examples/ActionSheet';
+import ToastExample from './examples/Toast';
 import SplitterExample from './examples/Splitter';
 import InputExample from './examples/Input';
 import IconExample from './examples/Icon';
@@ -140,6 +142,14 @@ class Examples extends React.Component {
       {
         title: 'Alert dialog',
         component: AlertDialogExample
+      },
+      {
+        title: 'Action sheet',
+        component: ActionSheetExample
+      },
+      {
+        title: 'Toast',
+        component: ToastExample
       },
       {
         title: 'Input',