+++ bindings/vue/examples/components/Dialogs.vue
@@ -9,6 +9,14 @@
       <v-ons-list-header>Notifications</v-ons-list-header>
       <v-ons-list-item
         tappable
+        @click="$ons.notification.toast('Hello, world!', {timeout: 2000})"
+      >
+        <div class="center">
+          Toast
+        </div>
+      </v-ons-list-item>
+      <v-ons-list-item
+        tappable
         @click="$ons.notification.alert('Hello, world!')"
       >
         <div class="center">
@@ -31,6 +39,14 @@
           Prompt
         </div>
       </v-ons-list-item>
+      <v-ons-list-item
+        tappable
+        @click="$ons.openActionSheet({buttons:['label1', 'label2', 'label3'], title: 'Lorem ipsum', cancelable: true, destructive: 1})"
+      >
+        <div class="center">
+          Action Sheet
+        </div>
+      </v-ons-list-item>
 
 
       <v-ons-list-header>Components</v-ons-list-header>
@@ -42,6 +58,16 @@
           Simple Dialog
         </div>
         <div class="right">Model: {{dialogVisible}}</div>
+      </v-ons-list-item>
+
+      <v-ons-list-item
+        tappable
+        @click="toastVisible = true"
+      >
+        <div class="center">
+          Toast
+        </div>
+        <div class="right">Model: {{toastVisible}}</div>
       </v-ons-list-item>
 
       <v-ons-list-item
@@ -83,6 +109,16 @@
         </div>
         <div class="right">Model: {{popoverVisible}}</div>
       </v-ons-list-item>
+
+      <v-ons-list-item
+        tappable
+        @click="actionSheetVisible = true"
+      >
+        <div class="center">
+          Action Sheet
+        </div>
+        <div class="right">Model: {{actionSheetVisible}}</div>
+      </v-ons-list-item>
     </v-ons-list>
 
     <v-ons-dialog cancelable
@@ -96,6 +132,8 @@
       Lorem ipsum
       <button @click="dialogVisible = !dialogVisible">toggle</button>
     </v-ons-dialog>
+
+    <v-ons-toast :visible="toastVisible"><div class="message">Cha-La Head-Cha-La</div><button @click="toastVisible = false">Dismiss</button></v-ons-toast>
 
     <v-ons-alert-dialog cancelable
       modifier="rowfooter"
@@ -142,6 +180,19 @@
       <button @click="popoverVisible = !popoverVisible">toggle</button>
     </v-ons-popover>
 
+    <v-ons-action-sheet
+      :visible="actionSheetVisible"
+      @update="actionSheetVisible = $event"
+      cancelable
+      title="Description"
+    >
+      <v-ons-action-sheet-button icon="md-square-o">Label 1</v-ons-action-sheet-button>
+      <v-ons-action-sheet-button icon="md-square-o">Label 2</v-ons-action-sheet-button>
+      <v-ons-action-sheet-button icon="md-square-o">Label 4</v-ons-action-sheet-button>
+      <v-ons-action-sheet-button icon="md-square-o" modifier="destructive">Label 4</v-ons-action-sheet-button>
+      <v-ons-action-sheet-button icon="md-square-o">Label 5</v-ons-action-sheet-button>
+    </v-ons-action-sheet>
+
   </v-ons-page>
 </template>
 
@@ -150,9 +201,11 @@
     data() {
       return {
         dialogVisible: false,
+        toastVisible: false,
         alertDialog1Visible: false,
         alertDialog2Visible: false,
         popoverVisible: false,
+        actionSheetVisible: false,
         modalVisible: false
       }
     },