+++ CHANGELOG.md
@@ -7,21 +7,49 @@ dev
 
 ### New Features
 
- * ons.createElement: Allows creating new elements from templates or inline HTML.
+ * css-components: Added `list-title` component.
+ * css-components: Added `action-sheet` component.
+ * core: Added `ons-list-title` element.
+ * core: Added `ons-action-sheet` element.
+ * core: Added `ons-action-sheet-button` element.
+ * core: Added new method `ons.actionSheet(...)` for creating inline `ons-action-sheet` elements.
+ * core: Added new method `ons.createElement(...)` that allows creating new elements from templates or inline HTML.
+ * core: A fake device back button event is now fired on ESC press.
+ * ons-splitter: Push animation.
+ * ons-splitter: Reveal animation.
+ * angular1: Added `ons-action-sheet` bindings.
  * core: Added `ons-card` element.
  * angular1: Added `ons-card` bindings.
 
 ### Bug Fixes
 
- * ons-input: Fixed [#1974](https://github.com/OnsenUI/OnsenUI/issues/1974).
- * ons-tabbar: Minor fix for situations where the '.page__content' is provided.
- * ons-splitter: Fixed timing for initial animations in bindings.
- * Device back button: Fixed possible crash in old platforms.
+ * angular1: Page loader now throws `destroy` event when page is unloaded.
+ * angular1: `myNavigator.topPage.data` should now be ready by the time the controller runs. Fixed [#1854](https://github.com/OnsenUI/OnsenUI/issues/1854).
+ * ons-carousel: Fixed [#1952](https://github.com/OnsenUI/OnsenUI/issues/1952).
+ * ons-dialog: Fixed mask-color attribute.
 
+ * ons.createElement: Allows creating new elements from templates or inline HTML.
+
+### Bug Fixes
+
+ * css-components: Fixed color of range component in Firefox. Fixed [#1964](https://github.com/OnsenUI/OnsenUI/issues/1964).
+ * ons-carousel: `refresh` event is now triggered when resized.
+ 
 ### BREAKING CHANGES
 
 * ons.createDialog, ons.createPopover, ons.createAlertDialog: Tags like `<ons-dialog>`, `<ons-alert-dialog>` or `<ons-popover>` are not added automatically anymore to the target template, they must be specified instead.
 * ons-template, external files: `ons-page` tag is not added automatically anymore as a wrapper of the target template. It must be manually specified.
+* ons.notification: Canceled notifications do not reject the returned promise anymore. Instead, when canceled they resolve to `-1` for `alert` and `confirm`, or `null` for `prompt`.
+
+v2.2.6
+----
+
+### Bug Fixes
+
+ * ons-splitter: Fixed timing for initial animations in bindings ([#1979](https://github.com/OnsenUI/OnsenUI/issues/1979), [#1985](https://github.com/OnsenUI/OnsenUI/issues/1985)).
+ * ons-tabbar: Minor fix for situations where the '.page__content' is provided ([#1978](https://github.com/OnsenUI/OnsenUI/issues/1978)).
+ * ons-input: Fixed [#1974](https://github.com/OnsenUI/OnsenUI/issues/1974).
+ * Device back button: Fixed possible crash in old platforms ([#1983](https://github.com/OnsenUI/OnsenUI/issues/1983)).
 
 v2.2.5
 ----