+++ bindings/react/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 dev
 ----
 * input: Added initial date value support.
+* action-sheet: Added new component.
+* toast: Implemented Toast component.
 * card: Added new component.
 
 v1.2.0